requirements.txt
gitpython==3.1.31
schedule==1.2.0
